requirement.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.7.1
django-tastypie==0.13.3
enum34==1.1.6
ipython==5.1.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
python-mimeparse==1.5.2
shell==1.0.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.0
wcwidth==0.1.7
wheel==0.24.0
